requirements.txt
Django==2.1.3
django-extensions==1.9.7
django-suit==0.2.25
django-stdimage==3.0.0
djangorestframework==3.7.3
olefile==0.44
Pillow==4.3.0
progressbar2==3.34.3
python-utils==2.2.0
pytz==2017.3
six==1.11.0
typing==3.6.2
django-robots==3.0
